+++ param_decomp/__init__.py
@@ -17,9 +17,9 @@ from param_decomp.configs import PDConfig
 from param_decomp.experiments.driver import ExperimentConfig, ExperimentDriver
 from param_decomp.load import load_pd
 from param_decomp.models.batch_and_loss_fns import PDTarget
-from param_decomp.pd_run import PDRun
 from param_decomp.run_metadata import RunMetadata
 from param_decomp.run_pd import run_pd
+from param_decomp.saved_run import PDRun
 
 __all__ = [
     "ExperimentConfig",
