+++ param_decomp/adapters/param_decomp.py
@@ -9,7 +9,7 @@ from param_decomp.autointerp.schemas import ModelMetadata
 from param_decomp.experiments.driver import ExperimentConfig
 from param_decomp.experiments.lm.experiment import LMExperimentConfig
 from param_decomp.models.component_model import ComponentModel
-from param_decomp.pd_run import PDRun
+from param_decomp.saved_run import PDRun
 from param_decomp.topology import TransformerTopology
 from param_decomp.utils.wandb_utils import parse_wandb_run_path
