+++ param_decomp/dataset_attributions/harvest.py
@@ -28,7 +28,7 @@ from param_decomp.experiments.lm.experiment import LMExperimentConfig
 from param_decomp.harvest.repo import HarvestRepo
 from param_decomp.log import logger
 from param_decomp.models.component_model import ComponentModel
-from param_decomp.pd_run import PDRun
+from param_decomp.saved_run import PDRun
 from param_decomp.topology import TransformerTopology, get_sources_by_target
 from param_decomp.utils.distributed_utils import get_device
 from param_decomp.utils.wandb_utils import parse_wandb_run_path
