+++ param_decomp/scripts/alpha_sweep/alpha_sweep.py
@@ -30,7 +30,7 @@ from param_decomp.experiments.lm.experiment import LMExperimentConfig
 from param_decomp.log import logger
 from param_decomp.models.component_model import ComponentModel
 from param_decomp.models.components import make_mask_infos
-from param_decomp.pd_run import PDRun
+from param_decomp.saved_run import PDRun
 from param_decomp.types import ModelPath
